requirements.txt
chardet==3.0.3
coreapi==2.3.1
coreschema==0.0.4
diff-match-patch==20121119
Django==1.11.1
django-braces==1.11.0
django-crispy-forms==1.6.1
django-debug-toolbar==1.8
django-extensions==1.7.9
django-filter==1.0.4
django-formtools==2.0
django-import-export==0.5.1
django-nested-admin==3.0.17
django-oauth-toolkit==0.12.0
django-rest-swagger==2.1.2
django-reversion==2.0.8
django-suit==0.2.25
djangorestframework==3.6.3
drf-generators==0.3.0
future==0.16.0
httplib2==0.9.2
itypes==1.1.0
Jinja2==2.9.6
MarkupSafe==1.0
oauthlib==2.0.1
olefile==0.44
openapi-codec==1.3.1
Pillow==4.1.1
psycopg2==2.7.1
python-monkey-business==1.0.0
pytz==2017.2
requests==2.14.2
simplejson==3.10.0
six==1.10.0
sqlparse==0.2.3
tablib==0.11.4
uritemplate==3.0.0
xadmin==0.6.1
XlsxWriter==0.9.6
